chowner fast refactoring
