+++ lib/languages/de.php
@@ -2027,7 +2027,6 @@ $translations = array(
     '' => '',
     '' => '',
     '' => '',
-    'chowner02' => 'Geocacher',
     'graph_statistics_01' => 'Statistike von www.{site_name} :: Datum',
     'graph_statistics_02' => 'Active caches',
     'graph_statistics_03' => 'Gefundene caches',
