+++ tpl/stdstyle/cacheAdoption/cacheAdoption.css
@@ -1,0 +1,43 @@
+
+
+#cachesOffered li,
+#cachesOwned li {
+    font-size:1.3em;
+}
+
+a:hover, a:visited, a:link, a:active {
+    text-decoration: none;
+}
+
+.message{
+    color: #3c763d;
+    background-color: #dff0d8;
+    border-color: #d6e9c6;
+
+    width: 749px;
+    text-align: center;
+    padding:15px;
+    margin-bottom: 10px;
+
+    border-radius: 15px;
+}
+
+.message-err{
+    color: #a94442;
+    background-color: #f2dede;
+    border-color: #ebccd1;
+}
+
+.close-but {
+    position: absolute;
+    right: 5%;
+}
+
+.alertMsg {
+    color: #ff0000;
+    padding: 20px 0 20px 0;
+}
+
+label {
+  font-weight: bold;
+}
