+++ tpl/stdstyle/cacheAdoption/cacheList.tpl.php
@@ -1,66 +1,5 @@
 
-<!-- this is temporary style definitions - I'll remove it soon -->
-<style>
-
-#cachesOffered li,
-#cachesOwned li {
-    font-size:1.3em;
-}
-
-a:hover, a:visited, a:link, a:active {
-    text-decoration: none;
-}
-
-.badge {
-    color: #3c763d;
-    background-color: #dff0d8;
-    border-color: #d6e9c6;
-
-    width: 92%;
-    text-align: center;
-    text-transform: uppercase;
-    font-weight: bold;
-    padding:1px 10px;
-    margin-bottom: 10px;
-
-    border-radius: 5px;
-}
-
-.accept {
-    color: green;
-}
-
-.abort, .refuse {
-    color: #a94442;
-    background-color: #f2dede;
-    border-color: #ebccd1;
-
-}
-
-.message{
-    color: #3c763d;
-    background-color: #dff0d8;
-    border-color: #d6e9c6;
-
-    width: 92%;
-    text-align: center;
-    padding:15px;
-    margin-bottom: 10px;
-
-    border-radius: 15px;
-}
-
-.message-err{
-    color: #a94442;
-    background-color: #f2dede;
-    border-color: #ebccd1;
-}
-
-.close-but {
-    position: absolute;
-    right: 5%;
-}
-</style>
+<link rel="stylesheet" href="<?=$view->cacheAdoption_css?>">
 
 <script type='text/javascript'>
     function reload(){
