--- tpl/stdstyle/chowner_chooseuser.tpl.php
@@ -1,50 +1,0 @@
-<font size="1">
-<table class="content" width="97%">
-    <colgroup>
-        <col width="200">
-        <col>
-    </colgroup>
-    <tr>
-      <td class="content2-pagetitle" colspan="2">
-        <img src="tpl/stdstyle/images/blue/email.png" class="icon32" align="middle" />
-        <b>{{adopt_04}} <a href="viewcache.php?cacheid={cacheid}"><font color="#ff0000">{cachename}</font></a></b>
-      </td>
-    </tr>
-    <tr><td colspan="2">&nbsp;</td></tr>
-    <tr>
-        <td>
-            <form action="chowner.php?action=addAdoptionOffer" method="post">
-                <table border='0' style="font-size: 120%;" width='97%'>
-                    <tr>
-                        <td colspan="2">
-                            {{adopt_05}}<br/><br/>
-                            <font color="#ff0000">{{adopt_06}}</font>
-                        </td>
-                    </tr>
-                    <tr>
-                        <td height="20" colspan="2">
-                        </td>
-                    </tr>
-                    <tr>
-                        <td width="40%">
-                            <b>{{adopt_07}}</b>
-                        </td>
-                        <td align="left" width="*">
-                            <input tabindex="1" type="text" size="30" name="username">
-                        </td>
-                    </tr>
-                    <tr>
-                        <td>
-                        </td>
-                        <td align="left">
-                            <input type="hidden" name ="cacheid" value="{cacheid}">
-                            <input type="submit" class="btn btn-primary" value="{{adopt_08}}">
-                        </td>
-                    </tr>
-                </table>
-            </form>
-        </td>
-    </tr>
-    <br/>
-</table>
-</font>
